Robin's new gait control code.
I've broken down the walking task a bit differently. The firmware on the
arduino takes data over serial in the form of (servo, position) pairs
and executes those motions at a speed which can be set up beforehand.
The actual walking is controlled by a python script called walk.py,
which sends the serial bytes to control the robot.