New branch for new table structure.
I'm experimenting with a gait table which is slightly higher-level.
Instead of listing the servos and position values, it will list the
limb, joint, and position names. The RoboCommander will translate these
before sending them over serial. This makes gait tables easier to read
and write as a common format.